requirements.txt
pysnmp
pandas
openpyxl
paho-mqtt==1.6.1
python-dotenv